// konfigurator-frontend/src/types/PdfProfile.ts
export interface PDFFormState {
  id: number;
  material: string;
  stärke: string;
  korrosionsschutz: string;
  gebHöhe: string;
  name: string;
  datum: string;
  artikelnr: string;
  au_nr: string;
  kunde: string;
  kom: string;
  blatt: string;
  data_uri: string;
  profiles: PDFProfile[];
  image: {
    url: string;
    public_id: string;
  };
}

export interface Pieces {
  chung: string;
  beze: string;
  amount: number;
  angle: string;
  bezeichnung: string;
  ending: string;
  id: number;
  length: number;
  pos: string;
  zuschnitt: string;
  bezeichung: string;
}

export interface PDFProfile {
  pos_nr_grave: boolean;
  t: boolean;
  zuschnitt: boolean;
  bereich: string;
  ja: any;
  getrennt: string;
  nachtrag: string;
  sonder: string;
  versatz: string;
  gebäudehöhe: string;
  nein2: any;
  ja2: any;
  c: any;
  b: any;
  nein: any;
  a: any;
  bis: string;
  zus: string;
  position: string;
  fur: string;
  au_nr: string;
  date: string;
  image: any;
  id: number;
  pieces_data: Pieces[];
  name: string;
  materialstärke: string;
  gefalle1: string;
  halterabst: number;
  formteil: string;
  uberstand: string;
  gefalle: string;
  lfm: string;
  abw: string;
  ai_ie: string;
  ai_ie1: string;
  ge_te: string;
  ge_te1: string;
  ak: string;
  ek: string;
  davon: string;
  stuck: string;
  ae_stuck: string;
  ie_stuck: string;
  a_mass: string;
  b_mass: string;
  c_mass: string;
  a_mass_angle: string;
  b_mass_angle: string;
  c_mass_angle: string;
  a_mass_overhang: string;
  c_mass_overhang: string;
  materialwerkstoff: string;
  korrosionsschutz: string;
  Gebäudehöhe: string;
  haltetyp: string;
  zusverb: string;
  zus_stuck: string;
  zus_davon: string;
  aetyp: string;
  ae_davon: string;
  ietyp: string;
  ie_davon: string;
  geftyp: string;
  gef_stuck: string;
  gef_davon: string;
  aktyp: string;
  ak_stuck: string;
  ak_davon: string;
  kunde: string;
  kom: string;
  blatt: string;
  data_uri: string;
}